feat : merombak total semua fitur guru dan admin yang berhubungan dengan jadwal

+++ database/migrations/2026_03_01_071510_student_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('nama', 30);
 
             $table->unsignedInteger('jurusan_id')->nullable();
-            $table->unsignedInteger('kelas_id')->nullable();
+            $table->unsignedBigInteger('kelas_id')->nullable();
 
             $table->string('foto_profil', 255)->nullable();
