+++ database/migrations/2026_03_27_090218_guru_mapel_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('guru_mapel', function(Blueprint $table){
             $table->id();
             $table->unsignedBigInteger('guru_id');
-            $table->unsignedInteger('mapel_id');
+            $table->unsignedBigInteger('mapel_id');
             $table->timestamps();
 
             $table->foreign('guru_id')
